WIP: Moving forms into the new interface
